Import Evennia default building includes CmdTeleport
